add handleMongooseError
