+++ controllers/contacts/getContactById.js
@@ -1,4 +1,4 @@
-const Contact = require("../../models/contact")
+const { Contact } = require("../../models/contact")
 const{HttpError} = require("../../helpers")
 
 const getContactById = async (req, res, next) => {
