+++ controllers/contacts/listContacts.js
@@ -1,4 +1,4 @@
-const Contact = require("../../models/contact")
+const { Contact } = require("../../models/contact")
 
 const listContacts = async (req, res, next) => {
   const result = await Contact.find()
