+++ controllers/contacts/removeContact.js
@@ -1,4 +1,4 @@
-const Contact = require("../../models/contact")
+const { Contact } = require("../../models/contact")
 const{HttpError} = require("../../helpers")
 
 const removeContact = async (req, res, next) => {
