+++ controllers/contacts/updateContact.js
@@ -1,4 +1,4 @@
-const Contact = require("../../models/contact")
+const { Contact } = require("../../models/contact")
 const { HttpError } = require("../../helpers")
 
 const updateContact = async (req, res, next) => {
