+++ helpers/handleMongooseError.js
@@ -1,0 +1,7 @@
+const handleMongooseError = (error, data, next) => {
+  const { name, code } = error;
+  error.status = name === "MongoServerError" && code === 11000 ? 409 : 400;
+  next();
+};
+
+module.exports = handleMongooseError
