+++ models/contact.js
@@ -1,22 +1,45 @@
-const { Schema, model } = require("mongoose")
+const { Schema, model } = require("mongoose");
+const { handleMongooseError } = require("../helpers");
+const Joi = require("joi");
 
 const contactSchema = new Schema({
-    name: {
-        type: String,
-        required: [true, 'Set name for contact'],
-    },
-    email: {
-        type: String,
-    },
-    phone: {
-        type: String,
-    },
-    favorite: {
-        type: Boolean,
-        default: false,
-    },
-})
+  name: {
+    type: String,
+    required: [true, "Set name for contact"],
+  },
+  email: {
+    type: String,
+  },
+  phone: {
+    type: String,
+  },
+  favorite: {
+    type: Boolean,
+    default: false,
+  },
+});
 
-const Contact = model("contact", contactSchema)
+contactSchema.post("save", handleMongooseError);
 
-module.exports = Contact;
+const addSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().required(),
+  phone: Joi.string().required(),
+  favorite: Joi.boolean(),
+});
+
+const updateSchema = Joi.object({
+  favorite: Joi.boolean().required(),
+});
+
+const schemas = {
+    addSchema,
+    updateSchema
+};
+
+const Contact = model("contact", contactSchema);
+
+module.exports = {
+    Contact,
+    schemas
+};
