+++ routes/api/contacts.js
@@ -3,7 +3,7 @@ const express = require('express')
 const ctrl = require("../../controllers/contacts")
 const { ctrlWrapper } = require("../../helpers")
 const { validateBody } = require("../../middlewares")
-const schemas = require("../../schemas/contacts")
+const { schemas } = require("../../models/contact")
 
 const router = express.Router()
